+++ crates/forge/src/scarb/config.rs
@@ -1,11 +1,7 @@
-use anyhow::{anyhow, bail, Result};
+use anyhow::Result;
 use cheatnet::runtime_extensions::forge_config_extension::config::BlockId;
-use itertools::Itertools;
-use serde::Deserialize;
-use std::{
-    collections::{HashMap, HashSet},
-    num::NonZeroU32,
-};
+use serde::{Deserialize, Deserializer};
+use std::{collections::HashSet, num::NonZeroU32};
 use url::Url;
 
 pub const SCARB_MANIFEST_TEMPLATE_CONTENT: &str = r#"
@@ -42,159 +38,63 @@ pub const SCARB_MANIFEST_TEMPLATE_CONTENT: &str = r#"
 
 #[allow(clippy::module_name_repetitions)]
 #[allow(clippy::struct_excessive_bools)]
-#[derive(Debug, PartialEq, Default)]
+#[derive(Debug, PartialEq, Default, Deserialize)]
 pub struct ForgeConfigFromScarb {
     /// Should runner exit after first failed test
+    #[serde(default)]
     pub exit_first: bool,
     /// How many runs should fuzzer execute
     pub fuzzer_runs: Option<NonZeroU32>,
     /// Seed to be used by fuzzer
     pub fuzzer_seed: Option<u64>,
     /// Display more detailed info about used resources
+    #[serde(default)]
     pub detailed_resources: bool,
     /// Save execution traces of all test which have passed and are not fuzz tests
+    #[serde(default)]
     pub save_trace_data: bool,
     /// Build profiles of all tests which have passed and are not fuzz tests
+    #[serde(default)]
     pub build_profile: bool,
     /// Generate a coverage report for the executed tests which have passed and are not fuzz tests
+    #[serde(default)]
     pub coverage: bool,
     /// Fork configuration profiles
+    #[serde(default, deserialize_with = "validate_forks")]
     pub fork: Vec<ForkTarget>,
     /// Limit of steps
     pub max_n_steps: Option<u32>,
 }
 
-#[non_exhaustive]
-#[derive(Debug, PartialEq, Clone)]
+#[derive(Debug, PartialEq, Clone, Deserialize)]
 pub struct ForkTarget {
     pub name: String,
     pub url: Url,
     pub block_id: BlockId,
 }
 
-impl ForkTarget {
-    pub fn new(name: &str, url: &str, block_id_type: &str, block_id_value: &str) -> Result<Self> {
-        let parsed_url = Url::parse(url).map_err(|_| anyhow!("Failed to parse fork url"))?;
-        let block_id = match block_id_type {
-            "number" => BlockId::BlockNumber(
-                block_id_value
-                    .parse()
-                    .map_err(|_| anyhow!("Failed to parse block number"))?,
-            ),
-            "hash" => BlockId::BlockHash(
-                block_id_value
-                    .parse()
-                    .map_err(|_| anyhow!("Failed to parse block hash"))?,
-            ),
-            "tag" => match block_id_value {
-                "latest" => BlockId::BlockTag,
-                _ => bail!("block_id.tag can only be equal to latest"),
-            },
-            block_id_key => bail!("block_id = {block_id_key} is not valid. Possible values are = \"number\", \"hash\" and \"tag\""),
-        };
+fn validate_forks<'de, D>(deserializer: D) -> Result<Vec<ForkTarget>, D::Error>
+where
+    D: Deserializer<'de>,
+{
+    // deserialize to Vec<ForkTarget>
+    let fork_targets = Vec::<ForkTarget>::deserialize(deserializer)?;
 
-        Ok(Self {
-            name: name.to_string(),
-            url: parsed_url,
-            block_id,
-        })
-    }
-}
-
-/// Represents forge config deserialized from Scarb.toml using basic types like String etc.
-#[allow(clippy::struct_excessive_bools)]
-#[derive(Deserialize, Debug, PartialEq, Default)]
-pub(crate) struct RawForgeConfig {
-    #[serde(default)]
-    /// Should runner exit after first failed test
-    pub exit_first: bool,
-    /// How many runs should fuzzer execute
-    pub fuzzer_runs: Option<NonZeroU32>,
-    /// Seed to be used by fuzzer
-    pub fuzzer_seed: Option<u64>,
-    #[serde(default)]
-    // Display more detailed info about used resources
-    pub detailed_resources: bool,
-    #[serde(default)]
-    /// Save execution traces of all test which have passed and are not fuzz tests
-    pub save_trace_data: bool,
-    #[serde(default)]
-    /// Build profiles of all tests which have passed and are not fuzz tests
-    pub build_profile: bool,
-    #[serde(default)]
-    /// Generate a coverage report for the executed tests which have passed and are not fuzz tests
-    pub coverage: bool,
-    #[serde(default)]
-    /// Fork configuration profiles
-    pub fork: Vec<RawForkTarget>,
-    /// Limit of steps
-    pub max_n_steps: Option<u32>,
-}
-
-#[derive(Deserialize, Debug, PartialEq, Default, Clone)]
-pub(crate) struct RawForkTarget {
-    pub name: String,
-    pub url: String,
-    pub block_id: HashMap<String, String>,
-}
-
-fn validate_raw_fork_config(raw_config: RawForgeConfig) -> Result<RawForgeConfig> {
-    let forks = &raw_config.fork;
-
-    let names: Vec<_> = forks.iter().map(|fork| &fork.name).collect();
+    let names: Vec<_> = fork_targets.iter().map(|fork| &fork.name).collect();
     let removed_duplicated_names: HashSet<_> = names.iter().collect();
-
     if names.len() != removed_duplicated_names.len() {
-        bail!("Some fork names are duplicated");
+        return Err(serde::de::Error::custom("Some fork names are duplicated"));
     }
 
-    forks
-        .iter()
-        .try_for_each(|fork| match fork.block_id.len() {
-            1 => Ok(()),
-            _ => bail!("block_id should be set once per fork"),
-        })?;
-
-    Ok(raw_config)
-}
-
-impl TryFrom<RawForgeConfig> for ForgeConfigFromScarb {
-    type Error = anyhow::Error;
-
-    fn try_from(value: RawForgeConfig) -> Result<Self, Self::Error> {
-        let value = validate_raw_fork_config(value)?;
-        let mut fork_targets = vec![];
-
-        for raw_fork_target in value.fork {
-            let (block_id_type, block_id_value) =
-                raw_fork_target.block_id.iter().exactly_one().unwrap();
-
-            fork_targets.push(ForkTarget::new(
-                raw_fork_target.name.as_str(),
-                raw_fork_target.url.as_str(),
-                block_id_type,
-                block_id_value,
-            )?);
-        }
-
-        Ok(ForgeConfigFromScarb {
-            exit_first: value.exit_first,
-            fuzzer_runs: value.fuzzer_runs,
-            fuzzer_seed: value.fuzzer_seed,
-            detailed_resources: value.detailed_resources,
-            save_trace_data: value.save_trace_data,
-            build_profile: value.build_profile,
-            coverage: value.coverage,
-            fork: fork_targets,
-            max_n_steps: value.max_n_steps,
-        })
-    }
+    Ok(fork_targets)
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
-    use num_bigint::BigInt;
+    use serde_json::json;
+    use starknet_types_core::felt::Felt;
+    use test_case::test_case;
     use url::Url;
 
     #[test]
@@ -204,7 +104,16 @@ mod tests {
         let block_id_type = "number";
         let block_id_value = "123";
 
-        let fork_target = ForkTarget::new(name, url, block_id_type, block_id_value).unwrap();
+        let json_str = json!({
+            "name": name,
+            "url": url,
+            "block_id": {
+                block_id_type: block_id_value
+            }
+        })
+        .to_string();
+
+        let fork_target = serde_json::from_str::<ForkTarget>(&json_str).unwrap();
 
         assert_eq!(fork_target.name, name);
         assert_eq!(fork_target.url, Url::parse(url).unwrap());
@@ -219,15 +128,22 @@ mod tests {
     fn test_fork_target_new_valid_hash() {
         let name = "TestFork";
         let url = "http://example.com";
-        let block_id_type = "hash";
-        let block_id_value = "0x1";
 
-        let fork_target = ForkTarget::new(name, url, block_id_type, block_id_value).unwrap();
+        let json_str = json!({
+            "name": name,
+            "url": url,
+            "block_id": {
+                "hash": "0x1"
+            }
+        })
+        .to_string();
+
+        let fork_target = serde_json::from_str::<ForkTarget>(&json_str).unwrap();
 
         assert_eq!(fork_target.name, name);
         assert_eq!(fork_target.url, Url::parse(url).unwrap());
         if let BlockId::BlockHash(hash) = fork_target.block_id {
-            assert_eq!(hash.to_bigint(), BigInt::from(1));
+            assert_eq!(hash, Felt::from_dec_str("1").unwrap());
         } else {
             panic!("Expected BlockId::BlockHash");
         }
@@ -237,10 +153,17 @@ mod tests {
     fn test_fork_target_new_valid_tag() {
         let name = "TestFork";
         let url = "http://example.com";
-        let block_id_type = "tag";
-        let block_id_value = "latest";
 
-        let fork_target = ForkTarget::new(name, url, block_id_type, block_id_value).unwrap();
+        let json_str = json!({
+            "name": name,
+            "url": url,
+            "block_id": {
+                "tag": "latest"
+            }
+        })
+        .to_string();
+
+        let fork_target = serde_json::from_str::<ForkTarget>(&json_str).unwrap();
 
         assert_eq!(fork_target.name, name);
         assert_eq!(fork_target.url, Url::parse(url).unwrap());
@@ -251,45 +174,43 @@ mod tests {
         }
     }
 
-    #[test]
-    fn test_fork_target_new_invalid_url() {
-        let name = "TestFork";
-        let url = "invalid_url";
-        let block_id_type = "number";
-        let block_id_value = "123";
-
-        let result = ForkTarget::new(name, url, block_id_type, block_id_value);
+    #[test_case(
+        &json!({
+            "name": "TestFork",
+            "url": "invalid_url",
+            "block_id": {
+                "number": "123"
+            }
+        }),
+        "relative URL without a base";
+        "invalid url"
+    )]
+    #[test_case(
+        &json!({
+            "name": "TestFork",
+            "url": "http://example.com",
+            "block_id": {
+                "number": "invalid_number"
+            }
+        }),
+        "invalid digit found in string";
+        "invalid number"
+    )]
+    #[test_case(
+        &json!({
+            "name": "TestFork",
+            "url": "http://example.com",
+            "block_id": {
+                "hash": "invalid_hash"
+            }
+        }),
+        "Failed to create Felt from string";
+        "invalid hash"
+    )]
+    fn test_fork_target_invalid_cases(input: &serde_json::Value, expected_error: &str) {
+        let json_str = input.to_string();
+        let result = serde_json::from_str::<ForkTarget>(&json_str);
         assert!(result.is_err());
-        assert_eq!(result.unwrap_err().to_string(), "Failed to parse fork url");
-    }
-
-    #[test]
-    fn test_fork_target_new_invalid_block_id_value_number() {
-        let name = "TestFork";
-        let url = "http://example.com";
-        let block_id_type = "number";
-        let block_id_value = "invalid_number";
-
-        let result = ForkTarget::new(name, url, block_id_type, block_id_value);
-        assert!(result.is_err());
-        assert_eq!(
-            result.unwrap_err().to_string(),
-            "Failed to parse block number"
-        );
-    }
-
-    #[test]
-    fn test_fork_target_new_invalid_block_id_value_hash() {
-        let name = "TestFork";
-        let url = "http://example.com";
-        let block_id_type = "hash";
-        let block_id_value = "invalid_hash";
-
-        let result = ForkTarget::new(name, url, block_id_type, block_id_value);
-        assert!(result.is_err());
-        assert_eq!(
-            result.unwrap_err().to_string(),
-            "Failed to parse block hash"
-        );
+        assert!(result.unwrap_err().to_string().contains(expected_error));
     }
 }