+++ crates/sncast/src/starknet_commands/account/mod.rs
@@ -11,8 +11,9 @@ use configuration::{
 };
 use serde_json::json;
 use sncast::{chain_id_to_network_name, decode_chain_id, helpers::configuration::CastConfig};
-use starknet::{core::types::Felt, signers::SigningKey};
-use std::{fmt, fs::OpenOptions, io::Write};
+use starknet::signers::SigningKey;
+use starknet_types_core::felt::Felt;
+use std::{collections::HashMap, fmt, fs::OpenOptions, io::Write};
 use toml::Value;
 
 pub mod create;
@@ -20,6 +21,8 @@ pub mod delete;
 pub mod deploy;
 pub mod import;
 pub mod list;
+
+type NestedMap<T> = HashMap<String, HashMap<String, T>>;
 
 #[derive(Args)]
 #[command(about = "Creates and deploys an account to the Starknet")]